Move Locality bound utils and views to Locality app
