+++ django_project/localities/utils.py
@@ -8,6 +8,7 @@ from datetime import datetime
 import requests
 
 from django.conf import settings
+from django.contrib.auth.models import User
 from django.contrib.gis.geos import Point, Polygon
 from django.contrib.gis.measure import D
 from django.core.serializers.json import DjangoJSONEncoder
@@ -16,11 +17,11 @@ from django.template import Context, Template
 
 from core.utilities import extract_time
 from localities.models import (
-    Attribute, Changeset, Country, Domain, Locality, LocalityArchive, Specification,
-    SynonymLocalities, UnconfirmedSynonym, User, Value, ValueArchive
+    Attribute, Changeset, Country, Domain, Locality, LocalityArchive,
+    Specification, SynonymLocalities, UnconfirmedSynonym, Value, ValueArchive
 )
 from localities.tasks import regenerate_cache, regenerate_cache_cluster
-from social_users.utils import get_profile, get_update_detail
+from social_users.utils import get_profile
 
 
 def get_what_3_words(geom):
@@ -562,3 +563,48 @@ def parse_bbox(bbox):
         raise ValueError
     # create polygon from bbox
     return Polygon.from_bbox(tmp_bbox)
+
+
+def user_updates(user, date):
+    updates = []
+    # from locality archive
+    ids = (
+        LocalityArchive.objects
+        .filter(changeset__social_user=user).filter(changeset__created__lt=date)
+        .order_by('-changeset__created')
+        .values('changeset', 'object_id')
+        .annotate(id=Min('id')).values('id')
+    )
+    updates_temp = (
+        LocalityArchive.objects
+        .filter(changeset__social_user=user).filter(changeset__created__lt=date)
+        .filter(id__in=ids)
+        .order_by('-changeset__created')
+        .values('changeset', 'changeset__created', 'changeset__social_user__username', 'version')
+        .annotate(edit_count=Count('changeset'), locality_id=Max('object_id'))[:10]
+    )
+    changesets = []
+    for update in updates_temp:
+        changesets.append(update['changeset'])
+        updates.append(get_update_detail(update))
+
+    # get from locality if not in Locality Archive yet
+    updates_temp = (
+        Locality.objects
+        .filter(changeset__social_user=user).exclude(changeset__in=changesets)
+        .order_by('-changeset__created')
+        .values('changeset', 'changeset__created', 'changeset__social_user__username', 'version')
+        .annotate(edit_count=Count('changeset'), locality_id=Max('id'))[:10]
+    )
+    for update in updates_temp:
+        updates.append(get_update_detail(update))
+
+    updates.sort(key=extract_time, reverse=True)
+    return updates[:10]
+
+
+def get_update_detail(update):
+    profile = get_profile(User.objects.get(username=update['changeset__social_user__username']))
+    update['nickname'] = profile.screen_name
+    update['changeset__created'] = update['changeset__created']
+    return update
