+++ django_project/social_users/views.py
@@ -1,20 +1,16 @@
 # -*- coding: utf-8 -*-
-import json
 import logging
-from datetime import datetime
 
 from django.contrib.auth import logout as auth_logout
 from django.contrib.auth.models import User
-from django.core.serializers.json import DjangoJSONEncoder
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
 from django.views.generic import TemplateView, View
 
 from braces.views import LoginRequiredMixin
 
-from localities.utils import extract_updates
 from social_users.models import Profile
-from social_users.utils import get_profile, user_updates
+from social_users.utils import get_profile
 
 LOG = logging.getLogger(__name__)
 
@@ -83,20 +79,3 @@ def save_profile(backend, user, response, *args, **kwargs):
         profile.profile_picture = url
     profile.save()
     return {'user': user}
-
-
-class GetUserUpdates(View):
-
-    def get(self, request):
-        date = request.GET.get('date')
-        user = request.GET.get('user')
-        if not date:
-            date = datetime.now()
-        user = get_object_or_404(User, username=user)
-        last_updates = user_updates(user, date)
-        updates = extract_updates(last_updates)
-        result = {}
-        result['last_update'] = updates
-        result = json.dumps(result, cls=DjangoJSONEncoder)
-
-        return HttpResponse(result, content_type='application/json')
